+++ src/configs/tailwind.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = {
+  extends: [
+    'stylelint-config-tailwindcss',
+    'stylelint-plugin-rhythmguard/configs/strict',
+  ],
+};