+++ src/index.js
@@ -15,5 +15,6 @@ module.exports.rules = {
 module.exports.configs = {
   recommended: require('./configs/recommended'),
   strict: require('./configs/strict'),
+  tailwind: require('./configs/tailwind'),
 };
 module.exports.presets = require('./presets');